Mitigate a leak of sender and receiver links in the session
Session was holding onto sender and receiver links forever even after
both ends had closed.  Adds tests and fixes to mitigate this under the
current API capabilities

+++ qpid-proton4j-engine/src/main/java/org/apache/qpid/proton4j/engine/impl/ProtonSenderState.java
@@ -126,8 +126,8 @@ public class ProtonSenderState implements ProtonLinkState<ProtonOutgoingDelivery
         int remoteDeliveryCount = (int) flow.getDeliveryCount();
         int newDeliveryCountLimit = remoteDeliveryCount + (int) flow.getLinkCredit();
 
-        long effectiveCredit = 0xFFFFFFFFL & (int)(newDeliveryCountLimit - existingDeliveryCount);
-        if ( effectiveCredit > 0 ) {
+        long effectiveCredit = 0xFFFFFFFFL & newDeliveryCountLimit - existingDeliveryCount;
+        if (effectiveCredit > 0) {
             creditState.updateCredit((int) effectiveCredit);
         } else {
             creditState.updateCredit(0);
